drivers: iio: sensors: Enable early buffer feature for SMI230, IAM20680 sensor
1. This patch allows all sensors driver mentioned above to buffer
accel and gyro samples from driver probe time till userspace
sensor service are up and running (or upto nth second).
Buffered samples can be retrieved at any time via
input subsystem by writing to below sysfs entry.
echo 1 > /sys/bus/iio/devices/iio:device%d/read_acc_boot_sample
echo 1 > /sys/bus/iio/devices/iio:device%d/read_gyro_boot_sample
Use config option to enable buffering feature.

2.Modified the some of API of sensor drivers to solve the compilation
for 5.15 kernel version.

+++ drivers/input/sensors/smi230/smi230_i2c_driver.c
@@ -148,7 +148,7 @@ static int smi230_acc_i2c_probe(struct i2c_client *client,
 
 	if ((smi230_i2c_adapter != NULL) &&
 			(smi230_i2c_adapter == client->adapter)) {
-		PINFO("%s i2c adapter is at %x", SENSOR_ACC_NAME, (unsigned int)client->adapter);
+		PINFO("%s i2c adapter is at %x", SENSOR_ACC_NAME, client->adapter);
 	}
 	else {
 		PERR("%s i2c driver is not initialized yet before ACC driver!", SENSOR_GYRO_NAME);
@@ -214,7 +214,7 @@ static int smi230_gyro_i2c_probe(struct i2c_client *client,
 
 	if (smi230_i2c_adapter == NULL) {
 		smi230_i2c_adapter = client->adapter;
-		PINFO("%s i2c adapter is at %x", SENSOR_GYRO_NAME, (unsigned int)client->adapter);
+		PINFO("%s i2c adapter is at %x", SENSOR_GYRO_NAME, client->adapter);
 	}
 	else {
 		PERR("%s i2c driver should be initialized first!", SENSOR_GYRO_NAME);
